Avoids overwriting of local project files
My local .idea folder was getting overwritten with the one from the repo. I have a Mac at home and so all my directories were getting switched to "C:\..." which is a Windows directory. The .idea files get regenerated any time the project is rebuild using Gradle, so it shouldn't be an issue to ignore these files.

http://stackoverflow.com/questions/11124053/accidentally-committed-idea-directory-files-into-git