Added automation to filling the database. any parameters we can get from yelp are now automatically added. This includes: Restaurant name, ID, Categories,Address,Phone, city, state.
NOTE: There is a while loop in the function that prevents yelp API calls unless the current geo location is known. This means that when testing the app on an emulator you will need to comment it out, as gps location doesn't seem to work on emulators at the moment.

Location of the while loop: Line 237